requirements.txt
absl-py==2.3.1
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
blinker==1.7.0
boto3==1.34.46
botocore==1.34.46
certifi==2023.11.17
cffi==2.0.0
chardet==5.2.0
click==8.1.6
cloud-init==25.2
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
contourpy==1.3.3
cryptography==41.0.7
cycler==0.12.1
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
flatbuffers==25.9.23
fonttools==4.60.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kiwisolver==1.4.9
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
mediapipe==0.10.18
ml_dtypes==0.5.3
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.0
pexpect==4.9.0
pillow==12.0.0
protobuf==4.25.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.23
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
s3transfer==0.10.1
scikit-learn==1.7.2
scipy==1.16.3
sentencepiece==0.2.1
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
sos==4.9.2
sounddevice==0.5.3
ssh-import-id==5.11
systemd-python==235
threadpoolctl==3.6.0
Twisted==24.3.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
wadllib==1.3.6
wheel==0.42.0
zope.interface==6.1
